perf(rust): finish integer-time migration — drop all hot-path f64
The f32 migration left absolute time as f64 for precision: monotonic
anchors (PID dt, state elapsed, tuner) and wall-clock (Status.timestamp,
scheduler). On the RP2350's M33 (single-precision FPU) every f64 op is
soft-float, so the control loop still paid for it each tick. This finishes
the deferred "integer-time rework", removing all f64 from the hot path.

Monotonic time -> i64 milliseconds (pid, state, tuner, controller):
- Anchors prev_time/start_time/last_update_time + tuner anchors become
  i64 ms; dt = (now_ms - prev) as f32 / 1000.0. Integer subtraction is
  exact (no cancellation) and never touches soft-float.
- Signed i64, not u64, on purpose: state's NTP-jump guard needs the
  negative delta (state_run_golden.csv has a backward 1097->1093 s jump);
  u64 saturating-sub would diverge. Controller drops mono_s = now_ms as
  f64/1000.0 and passes now_ms as i64.

Wall-clock -> i64 Unix seconds (controller, scheduler, protocol, csv,
json, server, api, firmware):
- Status.timestamp f64->i64; scheduler internals f64->i64 sec;
  RecoveryLog.elapsed_seconds f64->f32 (was a lossless f32->f64->f32
  round-trip); NtpClock::unix_seconds_f64 -> unix_seconds_i64;
  start_time_in_future(i64,i64). Seconds, not ms, because
  Command.start_time/ScheduledSnapshot are already u64 seconds and the
  CSV ISO column is second-granularity.
- /api/status JSON timestamp is now an integer; the web KilnStatus
  .timestamp field is unused, so this is web-safe. Sub-second wall-clock
  is dropped — only the scheduler consumed it (now fires at the second
  boundary, <=1 s earlier) and the unused JSON field.

SSR on-time -> f32 (ssr_schedule): the 10 Hz subtick computed
(duty/100 * cycle) in f64 for bit-exact int() truncation; now f32
(FPU-native). The golden fixture uses whole-% duties (f32-exact) and a
sub-ms rounding on a fractional duty is invisible at the ~100 ms relay
sub-tick.

Golden replays unchanged: discrete outputs (state, step index,
integral-freeze, relay on/off) match the f64 reference exactly; pid/
state/tuner replays feed (t*1000).round() as i64 ms (the 10 ms-resolution
jitter rounds cleanly). Host tests 165 green; firmware thumbv8m release +
clippy clean. Flash 747,032 -> 745,080 B (-1,952); RAM bss 202,460 ->
202,272 B (-188).

+++ rust/kiln-app/src/api.rs
@@ -122,7 +122,9 @@ pub fn schedule_fields_present(profile: Option<&str>, start_time: Option<f64>) -
 
 /// Whether a scheduled start is in the future — strict `>` so a start at exactly
 /// `now` is rejected, matching `start_time <= time.time()` (`handle_api_schedule`).
-pub fn start_time_in_future(start_time: f64, now: f64) -> bool {
+/// Both are Unix seconds (`i64`); the request's `start_time` is narrowed from the
+/// parsed JSON number at the call site.
+pub fn start_time_in_future(start_time: i64, now: i64) -> bool {
     start_time > now
 }
 
@@ -304,9 +306,9 @@ mod tests {
         assert!(!schedule_fields_present(Some("cone6"), Some(0.0)));
         assert!(!schedule_fields_present(None, Some(1.0)));
 
-        assert!(start_time_in_future(1000.0, 999.0));
-        assert!(!start_time_in_future(1000.0, 1000.0));
-        assert!(!start_time_in_future(1000.0, 1001.0));
+        assert!(start_time_in_future(1000, 999));
+        assert!(!start_time_in_future(1000, 1000));
+        assert!(!start_time_in_future(1000, 1001));
     }
 
     #[test]
